<Refact> : ProfileInfo 컴포넌트 서버통신 코드 분리
